requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html

astroid==2.4.2
certifi==2020.6.20
colorama==0.4.4
cycler==0.10.0
dask==2.30.0
decorator==4.4.2
fsspec==0.8.4
imageio==2.9.0
isort==5.6.4
joblib==0.17.0
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
locket==0.2.0
matplotlib==3.3.2
mccabe==0.6.1
networkx==2.5
numpy==1.19.4
pandas==1.1.4
partd==1.1.0
Pillow==8.0.1
pylint==2.6.0
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.3.1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
six==1.15.0
threadpoolctl==2.1.0
tifffile==2020.10.1
toml==0.10.2
toolz==0.11.1
torch==1.4.0+cpu
torchvision==0.5.0+cpu
typed-ast==1.4.1
wrapt==1.12.1
#starlette
kafka-python==2.0.2 
starlette==0.14.1
uvicorn==0.13.2
itsdangerous==1.1.0
aiofiles==0.6.0
Jinja2==2.11.1
httpx==0.11.1